Initial refactoring to support generalization to biofilms and mats
ReactionSedimentTransport is now split into four pieces:
  - ReactionSedimentGridn1D
  - ReactionSedimentPhys
  - ReactionSedimentTransportSolid
  - ReactionSedimentTransportSolute
