+++ docs/src/PALEOsediment_Reactions.md
@@ -6,6 +6,9 @@ CurrentModule = PALEOsediment.Sediment
 ```
 
 ```@docs
-SedimentTransport.ReactionSedimentTransport
+SedimentGrid.ReactionSedimentGridn1D
+SedimentPhys.ReactionSedimentPhys
 SedimentBioRates.ReactionSedimentBioRates
+SedimentTransport.ReactionSedimentTransportSolid
+SedimentTransport.ReactionSedimentTransportSolute
 ```
