+++ src/sediment/Sediment.jl
@@ -2,6 +2,10 @@ module Sediment
 
 import PALEOboxes as PB
 
+include("SedimentGrid.jl")
+
+include("SedimentPhys.jl")
+
 include("SedimentTransport.jl")
 
 include("SedimentBioRates.jl")
